requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy
h5py
Pillow
matplotlib
scikit-image
tqdm
pyiqa
